move registries pagination class to gwells (for general use)
