+++ app/gwells/pagination.py
@@ -1,0 +1,20 @@
+from collections import OrderedDict
+from rest_framework.pagination import LimitOffsetPagination
+from rest_framework.response import Response
+
+
+class APILimitOffsetPagination(LimitOffsetPagination):
+    """
+    Provides LimitOffsetPagination with custom parameters.
+    """
+
+    max_limit = 100
+
+    def get_paginated_response(self, data):
+        return Response(OrderedDict([
+            ('count', self.count),
+            ('next', self.get_next_link()),
+            ('previous', self.get_previous_link()),
+            ('offset', self.offset),
+            ('results', data)
+        ]))
